add: JobContextMetadata class and refactored OperatorImpl class
