+++ samza-core/src/main/java/org/apache/samza/context/TaskContextImpl.java
@@ -29,89 +29,69 @@ import org.apache.samza.system.SystemStreamPartition;
 import org.apache.samza.table.ReadWriteTable;
 import org.apache.samza.table.TableManager;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.function.Function;
 
 
 public class TaskContextImpl implements TaskContext {
-  private final TaskModel taskModel;
-  private final MetricsRegistry taskMetricsRegistry;
-  private final Function<String, KeyValueStore> keyValueStoreProvider;
-  private final TableManager tableManager;
-  private final CallbackScheduler callbackScheduler;
-  private final OffsetManager offsetManager;
-  private final JobModel jobModel;
-  private final StreamMetadataCache streamMetadataCache;
-  private final Map<String, Object> objectRegistry = new HashMap<>();
+    private final TaskModel taskModel;
+    private final MetricsRegistry taskMetricsRegistry;
+    private final Function<String, KeyValueStore> keyValueStoreProvider;
+    private final TableManager tableManager;
+    private final CallbackScheduler callbackScheduler;
+    private final OffsetManager offsetManager;
 
-  public TaskContextImpl(TaskModel taskModel,
-      MetricsRegistry taskMetricsRegistry,
-      Function<String, KeyValueStore> keyValueStoreProvider,
-      TableManager tableManager,
-      CallbackScheduler callbackScheduler,
-      OffsetManager offsetManager,
-      JobModel jobModel,
-      StreamMetadataCache streamMetadataCache) {
-    this.taskModel = taskModel;
-    this.taskMetricsRegistry = taskMetricsRegistry;
-    this.keyValueStoreProvider = keyValueStoreProvider;
-    this.tableManager = tableManager;
-    this.callbackScheduler = callbackScheduler;
-    this.offsetManager = offsetManager;
-    this.jobModel = jobModel;
-    this.streamMetadataCache = streamMetadataCache;
-  }
 
-  @Override
-  public TaskModel getTaskModel() {
-    return this.taskModel;
-  }
-
-  @Override
-  public MetricsRegistry getTaskMetricsRegistry() {
-    return this.taskMetricsRegistry;
-  }
-
-  @Override
-  public KeyValueStore getStore(String storeName) {
-    KeyValueStore store = this.keyValueStoreProvider.apply(storeName);
-    if (store == null) {
-      throw new IllegalArgumentException(String.format("No store found for storeName: %s", storeName));
+    public TaskContextImpl(TaskModel taskModel,
+                           MetricsRegistry taskMetricsRegistry,
+                           Function<String, KeyValueStore> keyValueStoreProvider,
+                           TableManager tableManager,
+                           CallbackScheduler callbackScheduler,
+                           OffsetManager offsetManager,
+                           JobModel jobModel,
+                           StreamMetadataCache streamMetadataCache) {
+        this.taskModel = taskModel;
+        this.taskMetricsRegistry = taskMetricsRegistry;
+        this.keyValueStoreProvider = keyValueStoreProvider;
+        this.tableManager = tableManager;
+        this.callbackScheduler = callbackScheduler;
+        this.offsetManager = offsetManager;
     }
-    return store;
-  }
 
-  @Override
-  public <K, V> ReadWriteTable<K, V> getTable(String tableId) {
-    return this.tableManager.getTable(tableId);
-  }
+    @Override
+    public TaskModel getTaskModel() {
+        return this.taskModel;
+    }
 
-  @Override
-  public CallbackScheduler getCallbackScheduler() {
-    return this.callbackScheduler;
-  }
+    @Override
+    public MetricsRegistry getTaskMetricsRegistry() {
+        return this.taskMetricsRegistry;
+    }
 
-  @Override
-  public void setStartingOffset(SystemStreamPartition systemStreamPartition, String offset) {
-    this.offsetManager.setStartingOffset(this.taskModel.getTaskName(), systemStreamPartition, offset);
-  }
+    @Override
+    public KeyValueStore getStore(String storeName) {
+        KeyValueStore store = this.keyValueStoreProvider.apply(storeName);
+        if (store == null) {
+            throw new IllegalArgumentException(String.format("No store found for storeName: %s", storeName));
+        }
+        return store;
+    }
 
-  // TODO SAMZA-1935: below methods are used by operator code; they should be decoupled from this client API
+    @Override
+    public <K, V> ReadWriteTable<K, V> getTable(String tableId) {
+        return this.tableManager.getTable(tableId);
+    }
 
-  public void registerObject(String name, Object value) {
-    this.objectRegistry.put(name, value);
-  }
+    @Override
+    public CallbackScheduler getCallbackScheduler() {
+        return this.callbackScheduler;
+    }
 
-  public Object fetchObject(String name) {
-    return this.objectRegistry.get(name);
-  }
+    @Override
+    public void setStartingOffset(SystemStreamPartition systemStreamPartition, String offset) {
+        this.offsetManager.setStartingOffset(this.taskModel.getTaskName(), systemStreamPartition, offset);
+    }
 
-  public JobModel getJobModel() {
-    return this.jobModel;
-  }
+    // TODO SAMZA-1935: below methods are used by operator code; they should be decoupled from this client API
 
-  public StreamMetadataCache getStreamMetadataCache() {
-    return this.streamMetadataCache;
-  }
+
 }
